producer/requirements.txt
kafka-python==2.0.2
requests==2.32.3
python-dotenv==1.0.1
SQLAlchemy==2.0.36
urllib3==2.2.3
pyodbc==5.2.0
schedule==1.2.2